requirements.txt
streamlit==1.31.0
PyMuPDF==1.23.8
numpy==1.26.2
scikit-learn==1.3.2
sentence-transformers==2.2.2
spacy==3.7.2
llama-cpp-python==0.2.38
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl